# Performance analysis and improvements