--- a/rebulk/match.py
+++ b/rebulk/match.py
@@ -5,6 +5,8 @@ Classes and functions related to matches
 """
 from collections import defaultdict, MutableSequence
 import copy
+import itertools
+
 try:
     from collections import OrderedDict  # pylint:disable=ungrouped-imports
 except ImportError:  # pragma: no cover
@@ -20,6 +22,7 @@ class MatchesDict(OrderedDict):
     """
     A custom dict with matches property.
     """
+
     def __init__(self):
         super(MatchesDict, self).__init__()
         self.matches = defaultdict(list)
@@ -33,18 +36,66 @@ class _BaseMatches(MutableSequence):
     _base = list
     _base_add = _base.append
     _base_remove = _base.remove
+    _base_extend = _base.extend
 
     def __init__(self, matches=None, input_string=None):
         self.input_string = input_string
         self._max_end = 0
         self._delegate = []
-        self._name_dict = defaultdict(_BaseMatches._base)
-        self._tag_dict = defaultdict(_BaseMatches._base)
-        self._start_dict = defaultdict(_BaseMatches._base)
-        self._end_dict = defaultdict(_BaseMatches._base)
-        self._index_dict = defaultdict(_BaseMatches._base)
+        self.__name_dict = None
+        self.__tag_dict = None
+        self.__start_dict = None
+        self.__end_dict = None
+        self.__index_dict = None
         if matches:
             self.extend(matches)
+
+    @property
+    def _name_dict(self):
+        if self.__name_dict is None:
+            self.__name_dict = defaultdict(_BaseMatches._base)
+            for name, values in itertools.groupby([m for m in self._delegate if m.name], lambda item: item.name):
+                _BaseMatches._base_extend(self.__name_dict[name], values)
+
+        return self.__name_dict
+
+    @property
+    def _start_dict(self):
+        if self.__start_dict is None:
+            self.__start_dict = defaultdict(_BaseMatches._base)
+            for start, values in itertools.groupby([m for m in self._delegate], lambda item: item.start):
+                _BaseMatches._base_extend(self.__start_dict[start], values)
+
+        return self.__start_dict
+
+    @property
+    def _end_dict(self):
+        if self.__end_dict is None:
+            self.__end_dict = defaultdict(_BaseMatches._base)
+            for start, values in itertools.groupby([m for m in self._delegate], lambda item: item.end):
+                _BaseMatches._base_extend(self.__end_dict[start], values)
+
+        return self.__end_dict
+
+    @property
+    def _tag_dict(self):
+        if self.__tag_dict is None:
+            self.__tag_dict = defaultdict(_BaseMatches._base)
+            for match in self._delegate:
+                for tag in match.tags:
+                    _BaseMatches._base_add(self.__tag_dict[tag], match)
+
+        return self.__tag_dict
+
+    @property
+    def _index_dict(self):
+        if self.__index_dict is None:
+            self.__index_dict = defaultdict(_BaseMatches._base)
+            for match in self._delegate:
+                for index in range(*match.span):
+                    _BaseMatches._base_add(self.__index_dict[index], match)
+
+        return self.__index_dict
 
     def _add_match(self, match):
         """
@@ -52,14 +103,19 @@ class _BaseMatches(MutableSequence):
         :param match:
         :type match: Match
         """
-        if match.name:
-            _BaseMatches._base_add(self._name_dict[match.name], (match))
-        for tag in match.tags:
-            _BaseMatches._base_add(self._tag_dict[tag], match)
-        _BaseMatches._base_add(self._start_dict[match.start], match)
-        _BaseMatches._base_add(self._end_dict[match.end], match)
-        for index in range(*match.span):
-            _BaseMatches._base_add(self._index_dict[index], match)
+        if self.__name_dict is not None:
+            if match.name:
+                _BaseMatches._base_add(self._name_dict[match.name], (match))
+        if self.__tag_dict is not None:
+            for tag in match.tags:
+                _BaseMatches._base_add(self._tag_dict[tag], match)
+        if self.__start_dict is not None:
+            _BaseMatches._base_add(self._start_dict[match.start], match)
+        if self.__end_dict is not None:
+            _BaseMatches._base_add(self._end_dict[match.end], match)
+        if self.__index_dict is not None:
+            for index in range(*match.span):
+                _BaseMatches._base_add(self._index_dict[index], match)
         if match.end > self._max_end:
             self._max_end = match.end
 
@@ -69,14 +125,19 @@ class _BaseMatches(MutableSequence):
         :param match:
         :type match: Match
         """
-        if match.name:
-            _BaseMatches._base_remove(self._name_dict[match.name], match)
-        for tag in match.tags:
-            _BaseMatches._base_remove(self._tag_dict[tag], match)
-        _BaseMatches._base_remove(self._start_dict[match.start], match)
-        _BaseMatches._base_remove(self._end_dict[match.end], match)
-        for index in range(*match.span):
-            _BaseMatches._base_remove(self._index_dict[index], match)
+        if self.__name_dict is not None:
+            if match.name:
+                _BaseMatches._base_remove(self._name_dict[match.name], match)
+        if self.__tag_dict is not None:
+            for tag in match.tags:
+                _BaseMatches._base_remove(self._tag_dict[tag], match)
+        if self.__start_dict is not None:
+            _BaseMatches._base_remove(self._start_dict[match.start], match)
+        if self.__end_dict is not None:
+            _BaseMatches._base_remove(self._end_dict[match.end], match)
+        if self.__index_dict is not None:
+            for index in range(*match.span):
+                _BaseMatches._base_remove(self._index_dict[index], match)
         if match.end >= self._max_end and not self._end_dict[match.end]:
             self._max_end = max(self._end_dict.keys())
 
@@ -311,7 +372,8 @@ class _BaseMatches(MutableSequence):
                     return rindex
         return self.max_end
 
-    def holes(self, start=0, end=None, formatter=None, ignore=None, seps=None, predicate=None, index=None):  # pylint: disable=too-many-branches,too-many-locals
+    def holes(self, start=0, end=None, formatter=None, ignore=None, seps=None, predicate=None,
+              index=None):  # pylint: disable=too-many-branches,too-many-locals
         """
         Retrieves a set of Match objects that are not defined in given range.
         :param start:
@@ -508,6 +570,7 @@ class Matches(_BaseMatches):
     """
     A custom list[Match] contains matches list.
     """
+
     def __init__(self, matches=None, input_string=None):
         self.markers = Markers(input_string=input_string)
         super(Matches, self).__init__(matches=matches, input_string=input_string)
@@ -521,6 +584,7 @@ class Markers(_BaseMatches):
     """
     A custom list[Match] containing markers list.
     """
+
     def __init__(self, matches=None, input_string=None):
         super(Markers, self).__init__(matches=None, input_string=input_string)
 
@@ -533,8 +597,9 @@ class Match(object):
     """
     Object storing values related to a single match
     """
+
     def __init__(self, start, end, value=None, name=None, tags=None, marker=None, parent=None, private=None,
-                 pattern=None, input_string=None, formatter=None, conflict_solver=None):
+                 pattern=None, input_string=None, formatter=None, conflict_solver=None, **kwargs):
         self.start = start
         self.end = end
         self.name = name
@@ -547,7 +612,7 @@ class Match(object):
         self.pattern = pattern
         self.private = private
         self.conflict_solver = conflict_solver
-        self.children = Matches([], input_string)
+        self._children = None
         self._raw_start = None
         self._raw_end = None
         self.defined_at = pattern.defined_at if pattern else defined_at()
@@ -558,6 +623,19 @@ class Match(object):
         2-tuple with start and end indices of the match
         """
         return self.start, self.end
+
+    @property
+    def children(self):
+        """
+        Children matches.
+        """
+        if self._children is None:
+            self._children = Matches(None, self.input_string)
+        return self._children
+
+    @children.setter
+    def children(self, value):
+        self._children = value
 
     @property
     def value(self):
@@ -736,13 +814,13 @@ class Match(object):
     def __eq__(self, other):
         if isinstance(other, Match):
             return self.span == other.span and self.value == other.value and self.name == other.name and \
-                self.parent == other.parent
+                   self.parent == other.parent
         return NotImplemented
 
     def __ne__(self, other):
         if isinstance(other, Match):
             return self.span != other.span or self.value != other.value or self.name != other.name or \
-                self.parent != other.parent
+                   self.parent != other.parent
         return NotImplemented
 
     def __lt__(self, other):