--- a/rebulk/pattern.py
+++ b/rebulk/pattern.py
@@ -25,7 +25,7 @@ class Pattern(object):
     def __init__(self, name=None, tags=None, formatter=None, value=None, validator=None, children=False, every=False,
                  private_parent=False, private_children=False, private=False, private_names=None, ignore_names=None,
                  marker=False, format_all=False, validate_all=False, disabled=lambda context: False, log_level=None,
-                 properties=None, post_processor=None):
+                 properties=None, post_processor=None, **kwargs):
         """
         :param name: Name of this pattern
         :type name: str
@@ -136,7 +136,7 @@ class Pattern(object):
         :return:
         :rtype:
         """
-        if len(match) < 0 or match.value == "":
+        if len(match) == 0 or match.value == "":
             return False
 
         pattern_value = get_first_defined(self.values, [match.name, '__parent__', None],
@@ -164,7 +164,7 @@ class Pattern(object):
         :return:
         :rtype:
         """
-        if len(child) < 0 or child.value == "":
+        if len(child) == 0 or child.value == "":
             return False
 
         pattern_value = get_first_defined(self.values, [child.name, '__children__', None],
@@ -243,7 +243,7 @@ class Pattern(object):
         :rtype:
         """
         if self.post_processor:
-            return call(self.post_processor, matches, self)
+            return self.post_processor(matches, self)
         return matches
 
     def _matches_privatize(self, matches):
@@ -335,7 +335,7 @@ class StringPattern(Pattern):
     """
 
     def __init__(self, *patterns, **kwargs):
-        call(super(StringPattern, self).__init__, **kwargs)
+        super(StringPattern, self).__init__(**kwargs)
         self._patterns = patterns
         self._kwargs = kwargs
         self._match_kwargs = filter_match_kwargs(kwargs)
@@ -349,9 +349,8 @@ class StringPattern(Pattern):
         return self._match_kwargs
 
     def _match(self, pattern, input_string, context=None):
-        for index in call(find_all, input_string, pattern, **self._kwargs):
-            yield call(Match, index, index + len(pattern), pattern=self, input_string=input_string,
-                       **self._match_kwargs)
+        for index in find_all(input_string, pattern, **self._kwargs):
+            yield Match(index, index + len(pattern), pattern=self, input_string=input_string, **self._match_kwargs)
 
 
 class RePattern(Pattern):
@@ -360,7 +359,7 @@ class RePattern(Pattern):
     """
 
     def __init__(self, *patterns, **kwargs):
-        call(super(RePattern, self).__init__, **kwargs)
+        super(RePattern, self).__init__(**kwargs)
         self.repeated_captures = REGEX_AVAILABLE
         if 'repeated_captures' in kwargs:
             self.repeated_captures = kwargs.get('repeated_captures')
@@ -403,21 +402,21 @@ class RePattern(Pattern):
         for match_object in pattern.finditer(input_string):
             start = match_object.start()
             end = match_object.end()
-            main_match = call(Match, start, end, pattern=self, input_string=input_string, **self._match_kwargs)
+            main_match = Match(start, end, pattern=self, input_string=input_string, **self._match_kwargs)
 
             if pattern.groups:
                 for i in range(1, pattern.groups + 1):
                     name = names.get(i, main_match.name)
                     if self.repeated_captures:
                         for start, end in match_object.spans(i):
-                            child_match = call(Match, start, end, name=name, parent=main_match, pattern=self,
-                                               input_string=input_string, **self._children_match_kwargs)
+                            child_match = Match(start, end, name=name, parent=main_match, pattern=self,
+                                                input_string=input_string, **self._children_match_kwargs)
                             main_match.children.append(child_match)
                     else:
                         start, end = match_object.span(i)
                         if start > -1 and end > -1:
-                            child_match = call(Match, start, end, name=name, parent=main_match, pattern=self,
-                                               input_string=input_string, **self._children_match_kwargs)
+                            child_match = Match(start, end, name=name, parent=main_match, pattern=self,
+                                                input_string=input_string, **self._children_match_kwargs)
                             main_match.children.append(child_match)
 
             yield main_match
@@ -429,7 +428,7 @@ class FunctionalPattern(Pattern):
     """
 
     def __init__(self, *patterns, **kwargs):
-        call(super(FunctionalPattern, self).__init__, **kwargs)
+        super(FunctionalPattern, self).__init__(**kwargs)
         self._patterns = patterns
         self._kwargs = kwargs
         self._match_kwargs = filter_match_kwargs(kwargs)
@@ -458,14 +457,14 @@ class FunctionalPattern(Pattern):
                     if self._match_kwargs:
                         options = self._match_kwargs.copy()
                         options.update(args)
-                    yield call(Match, pattern=self, input_string=input_string, **options)
+                    yield Match(pattern=self, input_string=input_string, **options)
                 else:
                     kwargs = self._match_kwargs
                     if isinstance(args[-1], dict):
                         kwargs = dict(kwargs)
                         kwargs.update(args[-1])
                         args = args[:-1]
-                    yield call(Match, *args, pattern=self, input_string=input_string, **kwargs)
+                    yield Match(*args, pattern=self, input_string=input_string, **kwargs)
 
 
 def filter_match_kwargs(kwargs, children=False):