--- a/rebulk/utils.py
+++ b/rebulk/utils.py
@@ -8,7 +8,7 @@ from collections import MutableSet
 from types import GeneratorType
 
 
-def find_all(string, sub, start=None, end=None, ignore_case=False):
+def find_all(string, sub, start=None, end=None, ignore_case=False, **kwargs):
     """
     Return all indices in string s where substring sub is
     found, such that sub is contained in the slice s[start:end].
@@ -65,10 +65,8 @@ def get_first_defined(data, keys, default_value=None):
     :rtype:
     """
     for key in keys:
-        try:
+        if key in data:
             return data[key]
-        except KeyError:
-            pass
     return default_value