--- a/rebulk/utils.py
+++ b/rebulk/utils.py
@@ -41,6 +41,7 @@ def find_all(string, sub, start=None, end=None, ignore_case=False, **kwargs):
     :return: all indices in the input string
     :rtype: __generator[str]
     """
+    _ = kwargs
     if ignore_case:
         sub = sub.lower()
         string = string.lower()