@Toilal 2016-11-27
Great !

Only one question before merging : Is the "Handle unused kwargs" commit really useful ?