@ratoaq2 2016-11-27
pylint was failing.
I had no time to check how to properly fix that.