@Toilal 2016-11-27
Ok I see ! I'll merge, but removing this commit add adding a pylint ignore comment instead. Thank you, it's a great job !