@Toilal 2016-11-27
On my computer, guessit unit tests runs 2 times faster on python 3.5.1, from 25s to 12.5s. Using python 2.7.11, the benefits are less revelant but still there from 15s to 12.5s.