@Toilal 2016-11-27
Released in rebulk 0.8.2